CT-77: More clean up with files and restructuring
