+++ app/web/src/assets/images/crush_it_logo.png
